add standalone module for stylesheets
