+++ app/src/components/routes/navigation/tabs/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import FontIcon from "react-native-vector-icons/FontAwesome5";
-import { colors } from "theme";
+import { colors } from "../../../../styles/theme";
 import {ClaimsNavigator,VendorsNavigator,SettingsNavigator} from "../stacks";
 
 const Tab = createBottomTabNavigator();
