+++ app/src/index.js
@@ -4,8 +4,8 @@ import { Provider } from "react-redux";
 import store from "./store";
 
 // Assets.
-import { imageAssets } from "theme/images";
-import { fontAssets } from "theme/fonts";
+import { imageAssets } from "./styles/theme/images";
+import { fontAssets } from "./styles/theme/fonts";
 import Router from "./components/routes";
 
 const App = () => {
