+++ app/src/styles/index.js
@@ -1,0 +1,5 @@
+﻿import list from "./list";
+
+export default {
+  list  
+};
